requirements.txt
Flask>=2.3
Flask.SQLAlchemy>=3.0
requests>=2.31.0
Pillow>=9.5.0
python.dotenv>=1.0.0
ttkbootstrap>=1.10.1
Flask-Mail>=0.9.1
itsdangerous>=2.1.2